requirements.txt
alembic==1.17.2
aniso8601==10.0.1
blinker==1.9.0
click==8.3.1
Flask==3.1.2
flask-marshmallow==1.3.0
Flask-Migrate==4.1.0
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
greenlet==3.3.0
iniconfig==2.3.0
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.3
marshmallow==4.1.2
marshmallow-sqlalchemy==1.4.2
packaging==25.0
pluggy==1.6.0
psycopg2-binary==2.9.11
Pygments==2.19.2
pytest==9.0.2
pytest-flask==1.3.0
python-dotenv==1.2.1
pytz==2025.2
six==1.17.0
SQLAlchemy==2.0.45
typing_extensions==4.15.0
Werkzeug==3.1.4
